Add first basic test
