+++ Cargo.toml
@@ -1,7 +1,8 @@
 [workspace]
 resolver = "2"
 members = [
-  "pkcs11"
+  "pkcs11",
+  "pkcs11/config_file",
 ]
 
 [profile.release]
