+++ pkcs11/config_file/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "config_file"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+log = "0.4.22"
+serde_yaml = "0.9.34"
+merge = { features = [
+  "derive",
+  "std",
+], default-features = false, version = "0.1.0" }
+hex = "0.4.3"
+thiserror = "1.0.63"
+
+[dev-dependencies]
+hex-literal = "0.4.1"
